Refactor: move json destructuring from DayZeroCase to Document #2714
